backend/pyproject.toml
[project]
name = "firstpr-backend"
version = "0.1.0"
description = "Backend for FirstPR - AI-powered repository analysis"
authors = [
    {name = "Your Name", email = "you@example.com"},
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "google-generativeai>=0.3.2",
    "redis>=5.0.1",
    "celery>=5.3.6",
    "SQLAlchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "tree-sitter>=0.20.4",
    "tree-sitter-languages>=1.10.2",
    "tenacity>=8.2.3",
    "python-multipart>=0.0.7",
    "flower>=2.0.1"
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --asyncio-mode=auto"
